InstrProf: XFAIL tests from r208940 on Linux
According to the buildbots, the new features for shared objects don't
work on ELF since it requires an -fPIC when building the profile
library.  XFAIL these tests for now.

It's possible we'll have to build two versions of the profile library on
Linux (one with -fPIC and one without), but it's not clear to me exactly
how to resolve this.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@208946 91177308-0d34-0410-b5e6-96231b3b80d8