compiler-rt: whitespace and uniformity for arm
Make the whitespace a bit more uniform in the various assembly routines.  This
also makes the assembly files a bit more uniform on the ARM side by explicitly
stating that it is using the unified syntax and that the contents of the code is
in the text section (or segment).  No functional change.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@209985 91177308-0d34-0410-b5e6-96231b3b80d8