builtins: re-enable building assembly sources
Re-apply SVN r213684 which was reverted in SVN r213724 since it broke the
build bots.  Add a tweak to enable inclusion of the assembly sources in
standalone build as well.

Original commit message:

This patch address the PR20360. The CMake assembler build system
ignores the .S assembly files in builtins library build. This patch
fixes the issue.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@214013 91177308-0d34-0410-b5e6-96231b3b80d8