log-path_test testcase - make the invalid path more invalid
The ASan test/asan/TestCases/log-path_test.cc testcase uses /INVALID as an invalid path and expects that the program will not be allowed to create or write to that file. This actually is a valid writable path on one of my setups. Let's make the path more invalid.

Reviewed at http://reviews.llvm.org/D6727



git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@224694 91177308-0d34-0410-b5e6-96231b3b80d8