tsan: don't initialize trace header in release mode
We are going to use only a small part of the trace with the default
value of history_size. However, the constructor writes to the whole trace.
It writes mostly zeros, so freshly mmaped memory will do.
The only non-zero field if mutex type used for debugging.

Reduces per-goroutine overhead by 8K.

https://code.google.com/p/thread-sanitizer/issues/detail?id=89



git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@229127 91177308-0d34-0410-b5e6-96231b3b80d8