[Sanitizers] Pass the correct arch to the symbolizer for x86_64h
I have no idea how to directly test that as it depends on a particular
(micro-)architecure of the host processor.
Combined with llvm's r240339 this should fix issues people might have
be seeing intermitently on Darwin haswell machines (the symbolizer
would use the wrong slice of the binary, thus potentially resolving
to the wrong symbol).

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@240379 91177308-0d34-0410-b5e6-96231b3b80d8