Revert "[CMake] [macho_embedded] Check to make sure the compiler supports the architectures before generating build targets"
This commit reverts r252525. I was not really thinking about this fix properly. This doesn't work because it relys on try_compile checks which do a full compile & link. I'm going to put in a temporary solution as an interm step until we have a way to perform compiler checks without linking.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@252569 91177308-0d34-0410-b5e6-96231b3b80d8