Move mmap_limit_mp test to Posix
This test isn't posix specific, but it doesn't pass on Windows and is
XFAILed. I suspect that this test, which is expected to fail, is causing
the hangs I'm seeing on our WinASan builder.  Moving it to Posix seems
to be the cleanest way to avoid running it on Windows.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@260480 91177308-0d34-0410-b5e6-96231b3b80d8