[libprofile] Adjust dummy condition to be more forgiving.
On PS4, we have to fake environment variables by passing extra command
line arguments, so the dummy test `argc > 1` was failing.
The condition is just a dummy condition that the compiler can't fold
away, so the number is arbitrary as long as the condition is false.

Increase the number it compares against.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@264491 91177308-0d34-0410-b5e6-96231b3b80d8