test: Use %clangxx in objc++ test files
These test in this change are objc++, but are built using %clang, not %clangxx.
The reason this works is the driver has been adding -lc++ for sanitizer enabled
builds. By making these tests use %clangxx, they no longer depend on the driver
linking to c++.  Doing so will allow us to prevent overlinking of libc++ for
applications.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@274989 91177308-0d34-0410-b5e6-96231b3b80d8