tsan: simplify meta mapping
Don't xor user address with kAppMemXor in meta mapping.
The only purpose of kAppMemXor is to raise shadow for ~0 user addresses,
so that they don't map to ~0 (which would cause overlap between
user memory and shadow).
For meta mapping we explicitly add kMetaShadowBeg offset,
so we don't need to additionally raise meta shadow.



git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@282403 91177308-0d34-0410-b5e6-96231b3b80d8