[Compiler-RT][Builtins] Remove XFAIL for mulsc3; NFC
It was XFAILed in r298974. However, the problem was not exposed on
the buildbot because hardfp flag was not passed during the test.

We can fix the CMAKE to pass the same flag as building the lib to the
RUN line to see if the problem is still there. For now, we remove the
XFAIL.

git-svn-id: http://llvm.org/svn/llvm-project/compiler-rt/trunk@298997 91177308-0d34-0410-b5e6-96231b3b80d8